+++ src/utils/getExceptionPathDetails.test.ts
@@ -32,7 +32,7 @@ describe("getExceptionPathDetails", () => {
     expect(result).toStrictEqual({
       field: "ArrestSummonsNumber",
       formattedFieldName: "Arrest summons number",
-      location: "Case information",
+      location: "Case Details",
       offenceOrderIndex: undefined,
       tab: "Case"
     })
